+++ django_boost/models/mixins.py
@@ -1,9 +1,11 @@
 import uuid
 
 from django.db import models
+from django.db import router
 from django.utils.timezone import now
 from django.utils.translation import gettext_lazy as _
 
+from django_boost.models.deletion import LogicalDeletionCollector
 from django_boost.models.manager import LogicalDeletionManager
 from django_boost.utils.functions import json_to_model, model_to_json
 
@@ -59,8 +61,15 @@ class LogicalDeletionMixin(models.Model):
     def delete(self, using=None, keep_parents=False, hard=False, deleted_at=None):
         if hard:
             return super().delete(using=using, keep_parents=keep_parents)
-        self.deleted_at = deleted_at if deleted_at is not None else self.get_deleted_value()
-        return self.save()
+        if self.pk is None:
+            raise ValueError(
+                "%s object can't be deleted because its %s attribute is set "
+                "to None." % (self._meta.object_name, self._meta.pk.attname)
+            )
+        using = using or router.db_for_write(self.__class__, instance=self)
+        collector = LogicalDeletionCollector(using=using, origin=self, deleted_at=deleted_at)
+        collector.collect([self], keep_parents=keep_parents)
+        return collector.delete()
 
     def revive(self, force_update=False, using=None):
         """Revive logical deleted item."""